Create folders for components
