+++ src/components/Card/Design/Palette.js
